+++ burrow/storm.go
@@ -8,19 +8,20 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
 
-package main
+package burrow
 
 import (
 	"encoding/json"
 	"errors"
-	log "github.com/cihub/seelog"
-	"github.com/linkedin/Burrow/protocol"
-	"github.com/samuel/go-zookeeper/zk"
 	"math/rand"
 	"regexp"
 	"strconv"
 	"sync"
 	"time"
+
+	log "github.com/cihub/seelog"
+	"github.com/prasincs/Burrow/protocol"
+	"github.com/samuel/go-zookeeper/zk"
 )
 
 type StormClient struct {
