+++ notifier/http_notifier.go
@@ -12,14 +12,15 @@ package notifier
 
 import (
 	"bytes"
-	log "github.com/cihub/seelog"
-	"github.com/linkedin/Burrow/protocol"
-	"github.com/pborman/uuid"
 	"io"
 	"io/ioutil"
 	"net/http"
 	"text/template"
 	"time"
+
+	log "github.com/cihub/seelog"
+	"github.com/pborman/uuid"
+	"github.com/prasincs/Burrow/protocol"
 )
 
 type HttpNotifier struct {
