+++ notifier/slack_notifier.go
@@ -15,11 +15,12 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	log "github.com/cihub/seelog"
-	"github.com/linkedin/Burrow/protocol"
 	"io/ioutil"
 	"net/http"
 	"time"
+
+	log "github.com/cihub/seelog"
+	"github.com/prasincs/Burrow/protocol"
 )
 
 type SlackNotifier struct {
